% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf.R
\name{extract.otf.image.3d}
\alias{extract.otf.image.3d}
\title{Title
Optimal Transport Allocation and Transport between 3D images}
\usage{
extract.otf.image.3d(
  from.image,
  to.image,
  mass.cost = 0,
  transport.type = 3,
  p.degree = 2
)
}
\arguments{
\item{from.image}{3d array}

\item{to.image}{3d array}

\item{mass.cost}{cost of allocating or removing mass}

\item{transport.type}{optimal transport mode \link{default=3}: \n
0 - Balanced (equalizes source and target mass)
1 - Add mass (Only allow addition of mass in soutrce or target)
2 - Subtract mass (Only allow subtraction of mass in source or target)
3 - Match source mass (Add or subtract in source to match target)
4 - Free (Add or subtract mass anywhere)
5 - Free source only (Add or subtract mass in source only)}

\item{p.degere}{optimal transport cost degree (i.e. cost^p.degreee)}
}
\value{
list with allocation and trasnport cost images in from and to
}
\description{
Compute optimal transport between X and Y and create allocation
and transport images
}
\examples{

from = array(runif(1000), dim=rep(10, 3))
to = array(runif(1000), dim=rep(10, 3))
otf = extract.otf.image.3d(from, to)
image(otf$difference.from[,,5])

}
