% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf.R
\name{extract.otf.points}
\alias{extract.otf.points}
\title{Title
Optimal Transport Allocation and Transport between Points}
\usage{
extract.otf.points(
  from,
  to,
  from.weight = 1,
  to.weight = 1,
  mass.cost = 0,
  transport.type = 3,
  p.degree = 2
)
}
\arguments{
\item{from}{3d array}

\item{to}{3d array}

\item{mass.cost}{cost of allocating or removing mass}

\item{transport.type}{optimal transport mode \link{default=3}: \n
0 - Balanced (equalizes source and target mass)
1 - Add mass (Only allow addition of mass in soutrce or target)
2 - Subtract mass (Only allow subtraction of mass in source or target)
3 - Match source mass (Add or subtract in source to match target)
4 - Free (Add or subtract mass anywhere)
5 - Free source only (Add or subtract mass in source only)}

\item{from.weights}{weights of from points}

\item{to.weights}{weights of to points}

\item{p.degere}{optimal transport cost degree (i.e. cost^p.degreee)}
}
\value{
list woth allocation and trasnport cost images in from and to
}
\description{
Compute optimal transport between X and Y and create allocation
and transport images
}
